+++ src/routes/(app)/eventos/+page.server.ts
@@ -9,15 +9,7 @@ import { kv } from "$lib/server/kv";
 
 // export const prerender = 'auto';
 export const load: PageServerLoad = async ({ parent, params, request, setHeaders }) => {
-
-  const cached = await kv.get("results");
-  if (cached) {
-    return {
-      results: cached,
-    };
-  }
-  
-  const results = await getSanityServerClient(false).fetch(resultsQuery);
+    const results = await getSanityServerClient(false).fetch(resultsQuery);
 
   if (!results) {
     throw error(500, "No se encontraron resultados");